vjtop  代码重构
